📝 chore(artist): update summary of Sigur Rós in _index.md
📝 chore(artist): update summary of Slowdive in _index.md

📝 chore(posts): add image for "Everything Is Alive" post

📝 feat(everything-is-alive-28131346): add new post "Everything Is Alive" by Slowdive

This commit adds a new post titled "Everything Is Alive" by Slowdive. The post includes information about the album, such as the title, artist, release date, and tracklisting. It also includes a description of the album and a track-by-track guide provided by Neil Halstead. The post features additional images, videos, and release information.

The purpose of this commit is to add a new post to the website to provide information and promote the album "Everything Is Alive" by Slowdive.

📝 feat: add new post "Tao Of The Dead" by ...And You Will Know Us By The Trail Of Dead

📝 feat: add new post "Átta" by Sigur Rós