📝 feat(artist): add Hawkwind artist page
📝 feat(artist): add Richard Wright artist page

📝 feat(artist): add Sparklehorse artist page

📝 feat(posts): add Aja post

📝 feat(posts): add Aja post images

📝 chore(index.md): add new post "Bird Machine" by Sparklehorse

📝 chore(index.md): add tracklisting and Apple Music link for "Bird Machine" by Sparklehorse

📝 chore(index.md): add videos and release images for "Bird Machine" by Sparklehorse

📝 chore(index.md): add release information for "Bird Machine" by Sparklehorse

📝 docs(get-the-message-the-best-of-electronic): add new post for "Get The Message The Best Of Electronic" album

This commit adds a new post for the album "Get The Message The Best Of Electronic" by Electronic. The post includes information about the album, such as the title, artist, release date, genres, and styles. It also includes a tracklisting and additional images of the album cover. The post provides a brief description of the album, highlighting the collaboration between Johnny Marr and Bernard Sumner and the fusion of dance music and Marr's guitar style. The post also includes release information, such as the format, label, catalog number, and a link to the Discogs page for more details.

📝 docs(metallic-spheres-in-colour-28431148): add metadata and tracklisting for "Metallic Spheres In Colour" by The Orb
🖼️ feat(metallic-spheres-in-colour-28431148): add cover image for "Metallic Spheres In Colour" by The Orb

📝 chore(rivers-of-heresy): add content for the "Rivers Of Heresy" album post

This commit adds the content for the "Rivers Of Heresy" album post. It includes the tracklisting, release information, Apple Music link, and images. The post also includes a description of the band Empire State Bastard and their musical influences. Each track on the album is discussed by the band members, providing insights into the lyrics and musical composition. The post also includes links to Wikipedia and Discogs for further information about the album.

📷 feat(posts): add images for "Space Ritual" and "The Harmony Codex" posts
📝 feat(posts): add content and metadata for "Space Ritual" and "The Harmony Codex" posts

📝 chore: add new album "Top Brass" by Matt Berry and "Wet Dream" by Richard Wright

📁 add(content/posts/top-brass-28421410): add new album "Top Brass" by Matt Berry
📁 add(content/posts/wet-dream-28427713): add new album "Wet Dream" by Richard Wright

🖼️ add(content/posts/top-brass-28421410): add cover image for "Top Brass" album
🖼️ add(content/posts/top-brass-28421410): add additional images for "Top Brass" album
🖼️ add(content/posts/wet-dream-28427713): add cover image for "Wet Dream" album
🖼️ add(content/posts/wet-dream-28427713): add additional images for "Wet Dream" album

📝 feat(content/posts/top-brass-28421410): add metadata and tracklisting for "Top Brass" album
📝 feat(content/posts/wet-dream-28427713): add metadata and tracklisting for "Wet Dream" album