feat: add new post "The Likes Of Us" by Big Big Train
This commit adds a new post titled "The Likes Of Us" by Big Big Train. The post includes information about the album, such as the title, artist, release date, and tracklisting. It also includes additional images related to the album. The post provides details about the album's genre and style, as well as a brief description of the album and its significance. The commit also includes the release information for the album, including the format, label, catalog number, and notes. Finally, the commit adds the album cover image for "The Likes Of Us" by Big Big Train.