feat: add Buzzcocks artist page and Singles Going Steady album page
Add a new artist page for Buzzcocks with the following details:
- Title: Buzzcocks
- Summary: A brief history of the band and its lineup changes over the years
- Image: buzzcocks.jpg
- Apple Music artist URL: https://music.apple.com/gb/artist/buzzcocks/468584
- Wikipedia URL: https://en.wikipedia.org/wiki/Buzzcocks

Add a new album page for Singles Going Steady with the following details:
- Title: Buzzcocks - Singles Going Steady
- Artist: Buzzcocks
- Album name: Singles Going Steady
- Date: 2024-03-12T12:43:37.000000Z
- Release date: 1979-09-25
- Release ID: 30026605
- Slug: singles-going-steady-30026605
- Hide summary: true
- Cover image: singles-going-steady-30026605.jpg
- Additional images: URLs of additional album images
- Genres: Rock
- Styles: Punk

Include a description of the album and its significance, as well as the tracklisting and Apple Music link. Also, include videos of selected songs from the album.

Release Information:
- Wikipedia URL: https://en.wikipedia.org/wiki/Singles_Going_Steady
- Format: Vinyl LP Compilation Numbered Reissue Remastered Repress (Orange Translucent)
- Label: Domino
- Catalog Number: REWIGLP129
- Notes: Description of the release and its features
- Discogs URL: [Buzzcocks - Singles Going Steady](https://www.discogs.com/release/30026605-Buzzcocks-Singles-Going-Steady)

feat(posts): add new post "The Collective" by Kim Gordon

This commit adds a new post titled "The Collective" by Kim Gordon. The post includes information about the album, such as the title, artist, release date, and tracklisting. It also includes additional images, videos, and release information. The post is now ready to be published.